requirements.txt
astroid==2.9.3
async-generator==1.10
attrs==21.4.0
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.0
colorama==0.4.4
cryptography==42.0.4
et-xmlfile==1.0.1
h11==0.12.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
numpy==1.22.2
openpyxl==3.0.9
outcome==1.1.0
pandas==1.4.1
pip==23.3
platformdirs==2.5.0
pycparser==2.21
pylint==2.12.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
selenium==4.15.1
setuptools==65.5.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.18
wheel==0.38.1
win-inet-pton==1.1.0
wrapt==1.13.3
wsproto==1.0.0
